computer graphics/opengl/requirements.txt
PyOpenGL
pygame
